+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Date format: DD/MM/YYYY
 - Added the `outfit_war_round` collection. This provides the same data as the `outfit_war_rounds` collection,
   but in a more easily queryable and joinable manner.
 - :warning: The `outfit_war_rounds` collection is now deprecated and will be removed after the end of Nexus Season 1.
+- Added the `outfit_war_match.round_id` field. Note that this will be null if the match does not conclusively fit within a single round.
 - Removed the `item.is_account_scoped` field. It did not provide any useful information, given that some items can be
   both character and account scoped depending on how they were purchased.